fixed calculation precision

+++ Server/MoneyManager/Model/Charts/Deposit/DepositPayment.cs
@@ -4,6 +4,6 @@
     {
         public string Name { get; set; }
 
-        public decimal Value { get; set; }
+        public float Value { get; set; }
     }
 }
